requirements.txt
# Core LLM Framework
langchain==0.3.0
langchain-groq==0.2.0
langchain-google-genai==2.0.0
langgraph==0.2.20

# API Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Frontend
streamlit==1.39.0

# Data Validation
pydantic==2.9.0
pydantic-settings==2.5.0

# Observability & Tracking
mlflow==2.16.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==24.1.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0

# Fuzzy Matching (for location validation)
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0